# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scraper/extract:
    post:
      operationId: extractMedia
      tags: [scraper]
      summary: Extract media from a supported social URL
      description: Validates a social media URL and forwards it to a configured media extraction provider when one is available.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaExtractInput"
      responses:
        "202":
          description: Extraction request accepted by the configured provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaExtractResponse"
        "400":
          description: Invalid or unsupported URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: No media extraction provider is configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /telegram/send:
    post:
      operationId: sendTelegramMessage
      tags: [integrations]
      summary: Send a Telegram alert
      description: Sends a prioritized alert through the configured Telegram bot and records the delivery event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramSendInput"
      responses:
        "200":
          description: Telegram alert delivered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramSendResponse"
        "400":
          description: Invalid Telegram alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Telegram integration is not configured or unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
    TelegramSendInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 4096
        priority:
          type: string
          enum: [normal, urgent]
          default: normal
        channels:
          type: array
          items:
            type: string
            enum: [telegram, discord]
          minItems: 1
          uniqueItems: true
          default: [telegram]
      required: [message]
    TelegramSendResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        priority:
          type: string
          enum: [normal, urgent]
        telegram_message_id:
          type: [number, "null"]
        discord_delivered:
          type: boolean
        log_id:
          type: string
      required: [success, message, priority, telegram_message_id, discord_delivered, log_id]
...
    MediaExtractInput:
      type: object
      properties:
        url:
          type: string
          maxLength: 4096
      required: [url]
    MediaExtractResponse:
      type: object
      properties:
        status:
          type: string
          enum: [accepted]
        message:
          type: string
        url:
          type: string
        platform:
          type: string
          enum: [tiktok, instagram, reddit, x]
        job_id:
          type: ["string", "null"]
      required: [status, message, url, platform, job_id]