# lib/api-zod/src/generated/types/telegramSendResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TelegramSendResponsePriority } from './telegramSendResponsePriority';

export interface TelegramSendResponse {
  success: boolean;
  message: string;
  priority: TelegramSendResponsePriority;
  /** @nullable */
  telegram_message_id: number | null;
  discord_delivered: boolean;
  log_id: string;
}